Polish git-gpg-verify (mostly by avoiding Bash word splitting)
While it technically works in this scenario, Bash word splitting is
confusing (why is my array of groups of keys acting like a plain array
of keys) and error-prone. It also causes Shellcheck to throw warnings.
Kicksecure doesn't succeed in avoiding word splitting universally, but
we generally aim to. The name of
`derivative_signing_key_fingerprint_list` was also confusing as the
array was no longer a list of keys, but a list of groups of related
keys. I renamed it to `derivative_signing_key_fingerprint_group_list` to
make it more clear what it is.