Split mount option into individual mount configuration options
The previous mount option mechanism in derivative-maker-docker-run
required one to remember the correct order of specifying volume mount
points, making it easy to mount a directory to the wrong location. Split
it into separate mount options to remedy this.

Also remove the path validation regex. It's easy to bypass, and would
prohibit the user from specifying certain legitimate path formats.