Implement --update-repo prototype in derivative-maker-docker
The previous derivative-maker-docker code always updated to the latest
code from upstream. Updating a repo is an inherently risky operation,
as it may cause data loss or compromise the repo, thus it needs to be
something the user explicitly requests to prevent surprises.

The implementation is not complete - the current way of checking out
submodules appears to also pull upstream submodule code, meaning that
you can only check out a git tag if you also update the repository. This
needs to be resolved.